requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
seaborn>=0.12.0
matplotlib>=3.7.0
openpyxl>=3.1.0
langchain>=0.1.0
langchain-community>=0.0.38
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
streamlit>=1.35.0
plotly>=6.3.0
